refactor: use sharedControls instead of controlOverrides
