+++ packages/superset-ui-demo/storybook/stories/plugins/plugin-chart-echarts/Funnel/constants.ts
